Refactor finance transaction handling by moving AdmissionTransaction, ExamTransaction, and MonthlyTransaction to a new Transaction namespace. Implement AddStudent and AddAdmissionFeeWithAddStudent actions for better separation of concerns. Update Login functionality to use a container component for improved state management and error handling. Remove deprecated code and enhance the user experience in the login process.
