+++ app/Http/Controllers/FinanceController.php
@@ -8,10 +8,7 @@ use App\Actions\Finance\Earning\AddAdmissionFee;
 use App\Actions\Finance\Earning\AddDueFee;
 use App\Actions\Finance\Earning\AddExamFee;
 use App\Actions\Finance\Earning\AddMonthlyFee;
-use App\Actions\Finance\Earning\AdmissionTransaction;
-use App\Actions\Finance\Earning\ExamTransaction;
 use App\Actions\Finance\Earning\MonthlyDiscount;
-use App\Actions\Finance\Earning\MonthlyTransaction;
 use App\Actions\Finance\Expense\AddOthersVoucher;
 use App\Actions\Finance\Expense\AddSalaryVoucher;
 use App\Actions\Finance\Expense\Expense;
@@ -20,6 +17,9 @@ use App\Actions\Finance\Paid\PaidList;
 use App\Actions\Finance\Reports\MonthlyDailyReport;
 use App\Actions\Finance\Reports\MonthlyGroupReport;
 use App\Actions\Finance\Summary;
+use App\Actions\Transaction\AdmissionTransaction;
+use App\Actions\Transaction\ExamTransaction;
+use App\Actions\Transaction\MonthlyTransaction;
 use App\Models\DailyReportApproval;
 use App\Models\Exam;
 use App\Models\ExpenseLog;
